+++ packages/firebase_auth/lib/src/firebase_user.dart
@@ -26,19 +26,21 @@ class FirebaseUser extends UserInfo {
   /// Returns true if the user's email is verified.
   bool get isEmailVerified => _data['isEmailVerified'];
 
-  /// Obtains the id token for the current user, forcing a [refresh] if desired.
+  /// Obtains the id token result for the current user, forcing a [refresh] if desired.
   ///
   /// Useful when authenticating against your own backend. Use our server
   /// SDKs or follow the official documentation to securely verify the
   /// integrity and validity of this token.
   ///
   /// Completes with an error if the user is signed out.
-  Future<String> getIdToken({bool refresh = false}) async {
-    return await FirebaseAuth.channel
-        .invokeMethod<String>('getIdToken', <String, dynamic>{
+  Future<IdTokenResult> getIdToken({bool refresh = false}) async {
+    final Map<String, dynamic> data = await FirebaseAuth.channel
+        .invokeMapMethod<String, dynamic>('getIdToken', <String, dynamic>{
       'refresh': refresh,
       'app': _app.name,
     });
+
+    return IdTokenResult(data, _app);
   }
 
   /// Associates a user account from a third-party identity provider with this