+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -25,6 +25,22 @@ const String kMockPhoneNumber = '5555555555';
 const String kMockVerificationId = '12345';
 const String kMockSmsCode = '123456';
 const String kMockLanguage = 'en';
+const String kMockIdTokenResultSignInProvider = 'password';
+const Map<dynamic, dynamic> kMockIdTokenResultClaims = <dynamic, dynamic>{
+  'claim1': 'value1',
+};
+const int kMockIdTokenResultExpirationTimestamp = 123456;
+const int kMockIdTokenResultAuthTimestamp = 1234567;
+const int kMockIdTokenResultIssuedAtTimestamp = 12345678;
+const Map<String, dynamic> kMockIdTokenResult = <String, dynamic>{
+  'token': kMockIdToken,
+  'expirationTimestamp': kMockIdTokenResultExpirationTimestamp,
+  'authTimestamp': kMockIdTokenResultAuthTimestamp,
+  'issuedAtTimestamp': kMockIdTokenResultIssuedAtTimestamp,
+  'signInProvider': kMockIdTokenResultSignInProvider,
+  'claims': kMockIdTokenResultClaims,
+};
+
 final int kMockCreationTimestamp = DateTime(2019, 1, 1).millisecondsSinceEpoch;
 final int kMockLastSignInTimestamp =
     DateTime.now().subtract(const Duration(days: 1)).millisecondsSinceEpoch;
@@ -65,7 +81,7 @@ void main() {
         log.add(call);
         switch (call.method) {
           case "getIdToken":
-            return kMockIdToken;
+            return kMockIdTokenResult;
             break;
           case "isSignInWithEmailLink":
             return true;
@@ -126,9 +142,28 @@ void main() {
     }
 
     test('getIdToken', () async {
+      void verifyIdTokenResult(IdTokenResult idTokenResult) {
+        expect(idTokenResult.token, equals(kMockIdToken));
+        expect(
+            idTokenResult.expirationTime,
+            equals(DateTime.fromMillisecondsSinceEpoch(
+                kMockIdTokenResultExpirationTimestamp * 1000)));
+        expect(
+            idTokenResult.authTime,
+            equals(DateTime.fromMillisecondsSinceEpoch(
+                kMockIdTokenResultAuthTimestamp * 1000)));
+        expect(
+            idTokenResult.issuedAtTime,
+            equals(DateTime.fromMillisecondsSinceEpoch(
+                kMockIdTokenResultIssuedAtTimestamp * 1000)));
+        expect(idTokenResult.signInProvider,
+            equals(kMockIdTokenResultSignInProvider));
+        expect(idTokenResult.claims, equals(kMockIdTokenResultClaims));
+      }
+
       final FirebaseUser user = await auth.currentUser();
-      expect(await user.getIdToken(), equals(kMockIdToken));
-      expect(await user.getIdToken(refresh: true), equals(kMockIdToken));
+      verifyIdTokenResult(await user.getIdToken());
+      verifyIdTokenResult(await user.getIdToken(refresh: true));
       expect(
         log,
         <Matcher>[